changes for authentication while place bid

+++ app/controllers/spree/orders_controller_decorator.rb
@@ -1,15 +1,19 @@
 module Spree
   OrdersController.class_eval do
     def populate
+      @prodval= JSON.parse params[:products]
+      variant_id = @prodval.first[1].to_i
+      unless spree_current_user
+        session['bid'] = Bid.new(:price => params[:bid_price], :variant_id =>  variant_id )
+        redirect_to spree.login_path
+        return
+      end
       if current_order.present?
         current_order.line_items.delete_all
       end
-      @prodval= JSON.parse params[:products]
-      variant_id = @prodval.first[1].to_i
-
-      bid = Bid.new(:price => params[:bid_price], :variant_id =>  variant_id, :user_id => spree_current_user.id )
-
-      if bid.save
+      @bid = Bid.find_or_initialize_by_variant_id_and_user_id(variant_id, spree_current_user.id)
+      @bid.price = params[:bid_price]
+      if @bid.save
         respond_to do |format|
           flash[:success] = "Your Bid placed successfully."
           format.html { redirect_to products_path }
